automatically determine number of EWSB and pole mass iterations
from the given precision goal.

Before, the number of EWSB and pole mass iterations were hard-coded
numbers.  Especially the number of EWSB iterations was quite large
(100), which lead to sometimes slowly converging points where there
exists no EWSB solution.

This commit improves the situation and leads to a speed up in cases
where the EWSB equations cannot be fulfilled.

The number of EWSB iterations is now set to

   -log10(ewsb_iteration_precision) * 10

and the number of mass iterations is set to

   std::abs(-log10(precision) * 10

Given the typical precision goals of 10^-4 or 10^-5, this leads to a
maximum number of iterations of 40 or 50, respectively, which are very
reasonable choices.