make test more stable
The test_CMSSM_QedQcd_no_convergence.sh test should check that the
determination of the SM(5) parameters does not fail.  So far, the
determination on all machines work for this point.  However, the test
seems to fail on some machines due to different results of the overall
iteration.

This commit explicitly checks if the determination of the SM(5)
parameters did succeed in the test.