Use the CPATH environment variable to search for headers
The `$CPATH` and `$CPLUS_INCLUDE_PATH` are environment variables which can be
used to indicate additional locations for headers (for example, if they are
installed in `~/.local/include`).

The commit also changes looks the Eigen headers in the `eigen3` subdirectory of
all default include paths (so for example, it will search in
`~/.local/include/eigen3`).

Lastly, this commit should support paths with spaces.  It does this by changing
the default separator to `:` in for loops.  This is done by setting the `IFS`
variable, though care must be taken to unset it afterwards (both within the
loop, and after the loop in case the for-loop has no elements).

Signed-off-by: JP-Ellis <josh@jpellis.me>