Attempt to deduce mass dimension of input parameters in simple cases
in order to simplify the arrived at form of the semi-analytic solutions.
To be safe, this is currently restricted to the very simple case of
settings of the form {<dimensionless model parameter>, p} => p is
dimensionless.

However, this might not be safe in general, e.g., in a model file
a dimensionful parameter is used to set a dimensionless parameter
(so that the fixed parameters are made to be numerically equal
without regard for the mass dimensions), in which case the above
would fail.  One option would be to try to detect inconsistent
settings, as would occur in the above case.  Alternatively, a
more robust approach would be to require the model file author
to specify the mass dimensions of the extra parameters.