Test for convergence in SUSY parameters as well as masses
and add some additional points to test at, based on the
benchmarks given in arXiv:1304.5462 and scans in the
multiple solution region in Figure 7 of that paper.

Currently, the default SM input parameters are used.
However, changing them to match Appendix A of the
above paper leads to worse convergence after a single
iteration of the two-scale solver.