Avoid usage of SARAH`SMQ[] as much as possible.
Reason: SARAH`SMQ[] requires that the list SARAH`SMParticles is filled
with all SM particles, including goldstone and Higgs bosons and
ghosts.  However, this correct filling is not ensured in SARAH so far.
For this reason we should use FlexibleSUSY's IsSMParticle[] function.
This function should do everything correctly in the future.