drop requirement of setting SMParticles
Since SARAH 4.11.0 the SMQ[] function takes the additional argument
"Higgs -> True|False", which can be used to include/exclude the
SM-like Higgs (iff it is a singlet).  Furthermore, Goldstone bosons
and ghosts are now correctly taken into account.

Thus, the setting of the SMParticles list (which was basically a
hack/workaround) is obsolete.  But, we now require SARAH 4.11.0 or
higher.