Bugfix: element-wise matrix product for constraints of the form A*B
In Mathematica A*B represents an element-wise product.  FS should
interpret expressions of this form in the same way.  Matrix products
should be written as MatMul[A,B] or A.B .

This fixes a bug in the MSSMatMGUT, where the GUT boundary condition
has the form (according to SLHA)

   T[Yu] = Auij * Yu

where Auij and Yu are matrices and Auij * Yu must be interpreted as
element-wise product.