clean-up mb calculation
The old conversion of mb(MS-bar) -> mb(DR-bar) contained a wrong 2L
conversion term O(as^2).  This wrong term was compensated by the
(optional) 2L-SQCD contributions to mb in the MSSM by Alexander
Bednyakov.

This commit cleans this mess up: Now the mb(MS-bar) -> mb(DR-bar)
conversion is correct and the inclusion of the 2L-SQCD contributions
is simpler.