correcting initial guess of SM(5) coupling constants
SOFTSUSY guesses the couplings of the SM(6) if Q > mt.  However, this
leads to a wrong guess in FS, where we assume that the QedQcd class
returns the couplings of the SM(5).  This commit fixes this issue.

Guess in SOFTSUSY (SM(6)):

 {alpha_1, alpha_2, alpha_3} = {0.0168674, 0.0345517, 0.107913}

Guess in FS (SM(5)) now:

 {alpha_1, alpha_2, alpha_3} = {0.0167817, 0.0351789, 0.107913}

The new initial guess in FS (SM(5)) is also closer to the true SM(5)
couplings (at MZ) after the iteration has converged:

 {alpha_1, alpha_2, alpha_3} = {0.0167694, 0.0350305, 0.1184}

So, the initial guess is now closer to the final SM(5) couplings,
which should improve convergence.