chop off small numerical numbers in the 3L RG-invariance test
Note: The 3L beta function for lambda (from SUSYHD) was taken from
Eq.(99) of [1307.3536].  The relevant term in this beta function is

   yt^4 (-50.201 g3^4)              (1)

The numerical value of this term contains Zeta[3] and further
constants, which match the 3L log^1 term of O(mt^2 yt^2 g3^4) to Mh^2
in the SM from Eq.(3.6) [1407.4336] (up to a normalization factor of
2/5).

I've checked that by varying the numeric value of (1), Mh^2 becomes
exactly RG invariant at the 3L level O(mt^2 yt^2 g3^4).  For example,
when I replace

   50201 -> 50200.85

in the 3L beta function of lambda, the RG invariance becomes close to
exact.